Move IdxRedirectResult class to client package. (#94)
This aligns better than in the DTO package.
